python_space/requirements.txt
# FastAPI Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.20

# TikTok & Video Download
yt-dlp>=2024.12.13
curl-cffi>=0.13.0  # Required for browser impersonation with TikTok
TikTokApi>=7.2.1  # For comment extraction
playwright>=1.56.0  # Required by TikTokApi

# Rate Limiting & Caching
slowapi>=0.1.9
python-dotenv>=1.0.0

# HTTP & Async
httpx>=0.27.0
aiofiles>=24.1.0

# Data validation
pydantic>=2.10.0
pydantic-settings>=2.7.0

# AI Services (Ollama - gratuito)
ollama>=0.3.0

# Image Generation
Pillow>=10.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=4.1.0
pytest-mock>=3.14.0

